# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AlternativePayload": {
        "properties": {
          "id": {
            "description": "Caller's identifier for this alternative",
            "title": "Id",
            "type": "string"
          },
          "levels": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Level keys for indicators declared on an ordered or unordered scale.",
            "title": "Levels",
            "type": "object"
          },
          "values": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "Numeric indicator values. Omit an indicator you have no value for; omission means unknown and is never imputed.",
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "id"
        ],
        "title": "AlternativePayload",
        "type": "object"
      },
      "Comparison": {
        "properties": {
          "comparisons": {
            "items": {
              "$ref": "#/components/schemas/HeadToHead"
            },
            "title": "Comparisons",
            "type": "array"
          },
          "leader": {
            "title": "Leader",
            "type": "string"
          },
          "leader_index": {
            "title": "Leader Index",
            "type": "integer"
          },
          "leader_score": {
            "title": "Leader Score",
            "type": "number"
          },
          "scores": {
            "items": {
              "type": "number"
            },
            "title": "Scores",
            "type": "array"
          },
          "wins": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "description": "Per alternative index, the indicators it is credited with winning on.",
            "title": "Wins",
            "type": "object"
          }
        },
        "required": [
          "wins",
          "leader",
          "leader_index",
          "leader_score",
          "scores",
          "comparisons"
        ],
        "title": "Comparison",
        "type": "object"
      },
      "Difference": {
        "properties": {
          "closes": {
            "title": "Closes",
            "type": "number"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "family": {
            "title": "Family",
            "type": "string"
          },
          "favours": {
            "title": "Favours",
            "type": "string"
          },
          "flips": {
            "title": "Flips",
            "type": "boolean"
          },
          "indicator": {
            "title": "Indicator",
            "type": "string"
          },
          "leader_value": {
            "title": "Leader Value",
            "type": "string"
          },
          "new_gap": {
            "title": "New Gap",
            "type": "number"
          },
          "rival_value": {
            "title": "Rival Value",
            "type": "string"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "required": [
          "indicator",
          "display_name",
          "family",
          "unit",
          "leader_value",
          "rival_value",
          "new_gap",
          "closes",
          "flips",
          "favours"
        ],
        "title": "Difference",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HeadToHead": {
        "properties": {
          "contributing": {
            "items": {
              "$ref": "#/components/schemas/Difference"
            },
            "title": "Contributing",
            "type": "array"
          },
          "decisive": {
            "items": {
              "$ref": "#/components/schemas/Difference"
            },
            "title": "Decisive",
            "type": "array"
          },
          "gap": {
            "title": "Gap",
            "type": "number"
          },
          "rival": {
            "title": "Rival",
            "type": "string"
          },
          "rival_index": {
            "title": "Rival Index",
            "type": "integer"
          },
          "rival_score": {
            "title": "Rival Score",
            "type": "number"
          },
          "rival_wins": {
            "items": {
              "$ref": "#/components/schemas/Difference"
            },
            "title": "Rival Wins",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "rival",
          "rival_index",
          "rival_score",
          "gap",
          "decisive",
          "contributing",
          "rival_wins",
          "summary"
        ],
        "title": "HeadToHead",
        "type": "object"
      },
...
      "ShortlistRequest": {
        "properties": {
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/AlternativePayload"
            },
            "maxItems": 5,
            "title": "Alternatives",
            "type": "array"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "context": {
            "items": {
              "type": "string"
            },
            "title": "Context",
            "type": "array"
          },
          "stakeholders": {
            "items": {
              "type": "string"
            },
            "title": "Stakeholders",
            "type": "array"
          }
        },
        "required": [
          "category"
        ],
        "title": "ShortlistRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Inference API.\n\nSubmit a shortlist and a context, receive a score per alternative. The response\nalways echoes the category's declared eligibility precondition, because the\nmodel ranks options that are assumed to have already passed regulatory\nprefiltering and saying so is the difference between a recommendation and a\ncompliance claim.\n\nThe registry travels inside the served model file, so the served semantics are\nexactly the ones the weights were trained under, with no database involved.\n",
    "title": "Context-adaptive product recommender",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/categories/{category_key}/indicators": {
      "get": {
        "operationId": "indicators",
        "parameters": [
          {
            "in": "path",
            "name": "category_key",
            "required": true,
            "schema": {
              "title": "Category Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Indicators",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Indicators"
      }
    },
    "/api/explore/compare": {
      "post": {
        "description": "Why the leading alternative is ahead of each of the others.",
        "operationId": "compare",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortlistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comparison"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare",
        "tags": [
          "explore"
        ]
      }
    },
    "/api/explore/context-sensitivity": {
      "post": {
        "operationId": "context_sensitivity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortlistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Context Sensitivity",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Context Sensitivity",
        "tags": [
          "explore"
        ]
      }
    },
...
    "/api/explore/response": {
      "get": {
        "operationId": "response",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "required": true,
            "schema": {
              "title": "Category",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "indicator",
            "required": true,
            "schema": {
              "title": "Indicator",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "context",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Context"
            }
          },
          {
            "in": "query",
            "name": "stakeholders",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stakeholders"
            }
          },
          {
            "in": "query",
            "name": "steps",
            "required": false,
            "schema": {
              "default": 21,
              "title": "Steps",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Response",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Response",
        "tags": [
          "explore"
        ]
      }
    },
    "/api/explore/stakeholder-sensitivity": {
      "post": {
        "operationId": "stakeholder_sensitivity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortlistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Stakeholder Sensitivity",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stakeholder Sensitivity",
        "tags": [
          "explore"
        ]
      }
    },
...
  }
}